Move AnalogStick component into its own file
